Removed resizeadapter to his own class
