+++ src/main/java/tudelft/ti2806/pl3/Application.java
@@ -2,6 +2,7 @@ package tudelft.ti2806.pl3;
 
 import newick.ParseException;
 import tudelft.ti2806.pl3.controls.KeyController;
+import tudelft.ti2806.pl3.controls.ResizeAdapter;
 import tudelft.ti2806.pl3.controls.ScrollListener;
 import tudelft.ti2806.pl3.controls.WindowController;
 import tudelft.ti2806.pl3.data.graph.GraphDataRepository;
@@ -23,8 +24,6 @@ import tudelft.ti2806.pl3.zoombar.ZoomBarController;
 import java.awt.Component;
 import java.awt.Dimension;
 import java.awt.Rectangle;
-import java.awt.event.ComponentAdapter;
-import java.awt.event.ComponentEvent;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
@@ -122,7 +121,7 @@ public class Application extends JFrame implements ControllerContainer {
 	}
 
 	private void setUpUi() {
-		this.addComponentListener(resizeAdapter());
+		this.addComponentListener(new ResizeAdapter(this).getResizer());
 		setZoomBarView();
 		setGraphView();
 		setSideBarView();
@@ -134,7 +133,7 @@ public class Application extends JFrame implements ControllerContainer {
 	public void makeGraphFromFolder() {
 		try {
 			File folder = FileSelector.selectFolder("Select data folder", this);
-			File[] files = FileSelector.getFilesFromFolder(folder, ".node.graph", ".edge.graph",".nwk");
+			File[] files = FileSelector.getFilesFromFolder(folder, ".node.graph", ".edge.graph", ".nwk");
 			makeGraph(files[0], files[1], files[2]);
 		} catch (FileSelectorException | NullPointerException exception) {
 			if (DialogUtil.confirm("Error!", "Your file was not found. Want to try again?")) {
@@ -277,35 +276,6 @@ public class Application extends JFrame implements ControllerContainer {
 		view.setVisible(true);
 	}
 
-	/**
-	 * Creates an adapter that updates screen sizes for the components in the view.
-	 *
-	 * @return the adapter
-	 */
-	private ComponentAdapter resizeAdapter() {
-		return new ComponentAdapter() {
-			@Override
-			public void componentResized(ComponentEvent e) {
-				Rectangle bounds = new Rectangle(main.getWidth(), main.getHeight());
-
-				size.setWidth((int) bounds.getWidth());
-				size.setHeight((int) bounds.getHeight());
-				size.calculate();
-
-				getSideBarController().getPanel().setBounds(0, size.getMenubarHeight(),
-						size.getSidebarWidth(), size.getHeight());
-				getGraphController().getPanel().setBounds(0, 0, size.getWidth(),
-						size.getHeight() - size.getZoombarHeight());
-				getZoomBarController().getPanel().setBounds(0,
-						size.getHeight() - size.getZoombarHeight(),
-						size.getWidth(), size.getZoombarHeight());
-				getPhyloController().getView().updateSize();
-
-				main.repaint();
-			}
-		};
-	}
-
 	@Override
 	public GraphController getGraphController() {
 		return graphController;
@@ -329,6 +299,14 @@ public class Application extends JFrame implements ControllerContainer {
 	@Override
 	public FindgenesController getFindgenesController() {
 		return findgenesController;
+	}
+
+	public Rectangle getBounds() {
+		return new Rectangle(main.getWidth(), main.getHeight());
+	}
+
+	public void repaint(){
+		main.repaint();
 	}
 
 }
